requirements.txt
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
PyPDF2
streamlit
pandas
python-dotenv
pydantic
numpy
